fashion360-eai/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.7.4
pydantic-settings==2.9.1
SQLAlchemy==2.0.31
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
python-jose==3.5.0
python-dotenv==1.1.0
email_validator==2.2.0
